+++ test/blog_test.exs
@@ -6,7 +6,7 @@ defmodule AshNeo4j.Blog.Test do
   @moduledoc false
   use ExUnit.Case, async: false
   alias AshNeo4j.Neo4jHelper
-  alias AshNeo4j.BoltxHelper
+  alias AshNeo4j.BoltyHelper
   alias AshNeo4j.Test.Resource.Author
   alias AshNeo4j.Test.Resource.Post
   alias AshNeo4j.Test.Resource.Comment
@@ -15,7 +15,7 @@ defmodule AshNeo4j.Blog.Test do
   import AshNeo4j.Test.Util, only: [check_enrichment: 5]
 
   setup_all do
-    BoltxHelper.start()
+    BoltyHelper.start()
   end
 
   setup do
@@ -27,9 +27,9 @@ defmodule AshNeo4j.Blog.Test do
     end)
   end
 
-  describe "Boltx configuration tests" do
+  describe "Bolty configuration tests" do
     test "neo4j is running" do
-      assert BoltxHelper.is_connected()
+      assert BoltyHelper.is_connected()
     end
   end